src/requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.64
langchain==0.1.14
pandas==1.5.3
python_dateutil==2.8.2
Requests==2.31.0
snowflake_connector_python==3.6.0
snowflake_snowpark_python==1.11.1
snowflake_sqlalchemy==1.4.7
